% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEfierWorkshop2024-pkg.R
\docType{package}
\name{iSEEfierWorkshop2024-pkg}
\alias{iSEEfierWorkshopSCBS2024}
\alias{iSEEfierWorkshopSCBS2024-package}
\alias{iSEEfierWorkshop2024-pkg}
\title{iSEEfierWorkshop2024: Learn how to use iSEE/iSEEfier to visualize omics data}
\description{
This is a Bioconductor Workshop package prapared for the Student Conference for Biological Sciences 2024 held in Tirana, Albania.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://najlaabassi.github.io/iSEEfierWorkshopSCBS2024/}
  \item Report bugs at \url{https://github.com/najlaabassi/iSEEfierWorkshopSCBS2024/issues/new}
}

}
\author{
\strong{Maintainer}: Najla Abassi \email{abassi.nejla96@gmail.com} (\href{https://orcid.org/0000-0001-8357-0938}{ORCID})

}
\keyword{internal}
