% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEfierWorkshop2024-data.R
\docType{data}
\name{sce_allen}
\alias{sce_allen}
\title{A sample \code{SingleCellExperiment} object}
\format{
A \code{SingleCellExperiment} object
}
\source{
Details on how this object has been created are included in the
\code{create_iSEEfierWorkshop2024_data.R} script, included in the (installed) \code{inst/scripts}
folder of the \code{iSEEfierWorkshop2024} package. This is also available at
\url{https://github.com/NajlaAbassi/iSEEfierWorkshop2024/devel/inst/scripts/create_iSEEfierWorkshop2024_data.R}
}
\description{
A sample \code{SingleCellExperiment} object
}
\details{
This \code{SingleCellExperiment} object is obtained from the \code{scRNAseq}
package. It corresponds to the \code{tophat_counts} assay of the \code{allen} dataset.

This object was processed to have log-transformed counts as well as different
dimentionality coordinates, including PCA, TSNE and UMAP.
}
\references{
Tasic et al. "Adult mouse cortical cell taxonomy revealed
by single cell transcriptomics", Nature Neuroscience, January 2016
doi:10.1038/nn.4216
}
